+++ src/modules/hellohook.c
@@ -47,7 +47,7 @@ void clientChangeCallback(RedisModuleCtx *ctx, RedisModuleEvent e, uint64_t sub,
     printf("Client %s event for client #%llu %s:%d\n",
         (sub == REDISMODULE_SUBEVENT_CLIENT_CHANGE_CONNECTED) ?
             "connection" : "disconnection",
-        ci->id,ci->addr,ci->port);
+        (unsigned long long)ci->id,ci->addr,ci->port);
 }
 
 void flushdbCallback(RedisModuleCtx *ctx, RedisModuleEvent e, uint64_t sub, void *data)
